feat(db): create tickets and stripe migrations

+++ config/migrations/3_create_tickets_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS tickets (
   id uuid primary key not null,
   name varchar(100) not null,
-  price decimal(7,2) not null,
+  price integer not null,
   total_qty integer not null,
   available_qty integer not null,
   description text,
